mod: re-apply migration add: title
