+++ src/app/layout.tsx
@@ -9,9 +9,11 @@ import { getMessages, getLocale } from 'next-intl/server';
 
 import './globals.css';
 import './scrollbar.css';
-import { DEFAULT_DESIGN_CONFIGURATION, DesignConfiguration  } from '@/db/types';
+import { DEFAULT_DESIGN_CONFIGURATION  } from '@/db/const';
 import { dbGetFederalStateByIdWithResult } from '@/db/functions/federal-state';
 import { getMaybeLogoFromS3 } from '@/s3';
+import { DesignConfiguration } from '@/db/types';
+
 const barlow = Barlow({
   weight: ['400', '500', '600', '700'],
   subsets: ['latin'],
@@ -19,11 +21,10 @@ const barlow = Barlow({
 
 export async function generateMetadata(): Promise<Metadata> {
   const maybeUser = await getMaybeUser();
-  console.log('maybeUser', maybeUser);
   const maybeLogoPath = await getMaybeLogoFromS3(maybeUser?.school.federalStateId);
-  console.log('maybeLogoPath', maybeLogoPath);
+  const [, federalState] = await dbGetFederalStateByIdWithResult(maybeUser?.school.federalStateId);
   return {
-    title: 'telli',
+    title: federalState?.telliName ?? 'telli',
     description: 'Der datenschutzkonforme KI-Chatbot für die Schule',
     icons: { icon: maybeLogoPath ?? '/telli.svg' },
   };
