+++ src/components/forms/attached-links.tsx
@@ -114,7 +114,7 @@ export function AttachedLinks({
             <div className="flex flex-row gap-2" key={`${field.id}-${index}`}>
               <Citation
                 source={field as unknown as WebsearchSource}
-                className="bg-secondary/40 rounded-enterprise-md h-10"
+                className="bg-secondary-dark rounded-enterprise-sm h-10"
                 handleDelete={() => handleDeleteLink(index)}
                 index={index}
                 sourceIndex={0}
